Several improvements.
Made the program NOT trigger when it couldn't connect to the PAX website and instead give textual feedback about it.
Made it so you can use Yahoo! or GMail for the email account. Added email and text address the program is using to the title of the program.
Added (unimplemented) auto-update -- need to make the downloaded version automatically run after it's finished downloading.
Changed test website to PAX main page.